* commit '981358db08919af6875d15620b51789e0256ed1d':
  CLA-2638 - Adjusted size property from INT to LONG

+++ Sentry.data.Infrastructure/ServiceImplementations/DataProcessing/DataFlowProvider.cs
@@ -130,7 +130,7 @@ namespace Sentry.data.Infrastructure
                     Logger.AddContextVariable(new TextVariable("flowexecutionguid", stepEvent.FlowExecutionGuid));
 
                     S3ObjectEvent s3Event = JsonConvert.DeserializeObject<S3ObjectEvent>(stepEvent.OriginalS3Event);
-                    Logger.AddContextVariable(new LongVariable("objectsize", (long)s3Event.s3.Object.size));
+                    Logger.AddContextVariable(new LongVariable("objectsize", s3Event.s3.Object.size));
                     if (stepEvent.RunInstanceGuid != null)
                     {
                         Logger.AddContextVariable(new TextVariable("runinstanceguid", stepEvent.RunInstanceGuid));
